Fixed dependency issues when saving a purchase

+++ src/main/java/ca/mcgill/ecse321/gameshop/DAO/RefundRequestRepository.java
@@ -4,5 +4,5 @@ import ca.mcgill.ecse321.gameshop.persistence.RefundRequest;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RefundRequestRepository extends CrudRepository<RefundRequest, Long> {
-    RefundRequest findRefundRequestById(Long id);
+    RefundRequest findRefundRequestById(Integer id);
 }
